fix: 모놀리식 서버로 구성 변경
