+++ order/src/main/java/com/klp/logistics/LogisticsApplication.java
@@ -1,0 +1,13 @@
+package com.klp.logistics;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LogisticsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(LogisticsApplication.class, args);
+    }
+
+}
