+++ order/src/main/java/com/klp/logistics/order/application/service/dto/OrderResponse.java
@@ -1,0 +1,6 @@
+package com.klp.logistics.order.application.service.dto;
+
+public record OrderResponse(
+
+) {
+}
