+++ order/src/main/java/com/klp/logistics/order/domain/entity/idempotencykey/OperationType.java
@@ -1,0 +1,14 @@
+package com.klp.logistics.order.domain.entity.idempotencykey;
+
+public enum OperationType {
+    INCREASE("증감"),
+    DECREASE("감소"),
+    MAKING("생성"),
+    CANCEL("취소");
+
+    private String description;
+
+    private OperationType(String description) {
+        this.description = description;
+    }
+}
