+++ order/src/main/java/com/klp/logistics/order/domain/entity/idempotencykey/Target.java
@@ -1,0 +1,12 @@
+package com.klp.logistics.order.domain.entity.idempotencykey;
+
+public enum Target {
+    DELIVERY("배송"),
+    INVENTORY("재고");
+
+    private String description;
+
+    Target(String description) {
+        this.description = description;
+    }
+}
