+++ order/src/main/java/com/klp/logistics/order/domain/entity/orderitem/OrderItemCommand.java
@@ -1,0 +1,10 @@
+package com.klp.logistics.order.domain.entity.orderitem;
+
+import java.util.UUID;
+
+public record OrderItemCommand(
+    UUID productId,
+    int quantity
+) {
+
+}
