+++ order/src/main/java/com/klp/logistics/order/global/config/AuditConfig.java
@@ -1,0 +1,10 @@
+package com.klp.logistics.order.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class AuditConfig {
+
+}
