--- order/src/main/java/com/klp/order/OrderApplication.java
@@ -1,13 +1,0 @@
-package com.klp.order;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrderApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(OrderApplication.class, args);
-    }
-
-}
