--- order/src/main/java/com/klp/order/application/service/dto/OrderResponse.java
@@ -1,6 +1,0 @@
-package com.klp.order.application.service.dto;
-
-public record OrderResponse(
-
-) {
-}
