--- order/src/main/java/com/klp/order/domain/entity/idempotencykey/Target.java
@@ -1,12 +1,0 @@
-package com.klp.order.domain.entity.idempotencykey;
-
-public enum Target {
-    DELIVERY("배송"),
-    INVENTORY("재고");
-
-    private String description;
-
-    Target(String description) {
-        this.description = description;
-    }
-}
