--- order/src/main/java/com/klp/order/domain/entity/orderitem/OrderItemCommand.java
@@ -1,10 +1,0 @@
-package com.klp.order.domain.entity.orderitem;
-
-import java.util.UUID;
-
-public record OrderItemCommand(
-    UUID productId,
-    int quantity
-) {
-
-}
