--- order/src/main/java/com/klp/order/global/config/AuditConfig.java
@@ -1,10 +1,0 @@
-package com.klp.order.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class AuditConfig {
-
-}
