--- order/src/main/java/com/klp/order/infrastructure/repository/OrderRepository.java
@@ -1,8 +1,0 @@
-package com.klp.order.infrastructure.repository;
-
-import com.klp.order.domain.entity.order.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-
-}
