--- order/src/test/java/com/klp/order/domain/entity/orderitem/OrderItemTest.java
@@ -1,130 +1,0 @@
-package com.klp.order.domain.entity.orderitem;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
-
-import com.klp.order.domain.entity.order.Order;
-import java.util.List;
-import java.util.UUID;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-@DisplayName("OrderItem 엔티티 테스트")
-public class OrderItemTest {
-
-    private UUID productId;
-    private int quantity;
-    private Order order;
-
-    @BeforeEach
-    void setUp() {
-        productId = UUID.randomUUID();
-        quantity = 10;
-        List<OrderItemCommand> initialItems = List.of(
-            new OrderItemCommand(UUID.randomUUID(), 1)
-        );
-        order = Order.create(1L, 2L, "테스트 주문", initialItems);
-    }
-
-    @Test
-    @DisplayName("주문 아이템을 생성할 수 있다")
-    void createOrderItem_Success() {
-        // given
-        OrderItemCommand command = new OrderItemCommand(productId, quantity);
-
-        // when
-        OrderItem orderItem = OrderItem.of(order, command);
-
-        // then
-        assertThat(orderItem.getProductId()).isEqualTo(productId);
-        assertThat(orderItem.getQuantity()).isEqualTo(quantity);
-        assertThat(orderItem.getOrder()).isEqualTo(order);
-        assertThat(orderItem.getDeliveryId()).isNull();
-    }
-
-    @Test
-    @DisplayName("주문 아이템 생성시 상품ID 가 Null 이면 예외가 발생한다")
-    void createOrderItem_Fail_ProductId_is_Null() {
-        // given
-        OrderItemCommand command = new OrderItemCommand(null, quantity);
-
-        // when & then
-        assertThatThrownBy(() -> OrderItem.of(order, command))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("상품 ID는 필수입니다.");
-    }
-
-    @Test
-    @DisplayName("주문 아이템 생성시 수량이 0 이하면 예외가 발생한다")
-    void createOrderItem_Fail_Quantity_is_Zero_or_Negative() {
-        // when & then
-        assertThatThrownBy(() -> OrderItem.of(order, new OrderItemCommand(productId, 0)))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("주문 수량은 1개 이상이어야 합니다.");
-
-        assertThatThrownBy(() -> OrderItem.of(order, new OrderItemCommand(productId, -5)))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("주문 수량은 1개 이상이어야 합니다.");
-    }
-
-    @Test
-    @DisplayName("주문 아이템 생성시 Order가 Null이면 예외가 발생한다")
-    void createOrderItem_Fail_Order_is_Null() {
-        // given
-        OrderItemCommand command = new OrderItemCommand(productId, quantity);
-
-        // when & then
-        assertThatThrownBy(() -> OrderItem.of(null, command))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("주문은 필수입니다.");
-    }
-
-    @Test
-    @DisplayName("주문에 배송ID 를 할당할 수 있다")
-    void assignDeliveryId_Success() {
-        // given
-        OrderItemCommand command = new OrderItemCommand(productId, quantity);
-        OrderItem orderItem = OrderItem.of(order, command);
-        UUID deliveryId = UUID.randomUUID();
-
-        // when
-        orderItem.assignDeliveryId(deliveryId);
-
-        // then
-        assertThat(orderItem.getDeliveryId()).isEqualTo(deliveryId);
-    }
-
-    @Test
-    @DisplayName("주문 수량을 수정할 수 있다")
-    void updateQuantity_Success() {
-        // given
-        OrderItemCommand command = new OrderItemCommand(productId, quantity);
-        OrderItem orderItem = OrderItem.of(order, command);
-        int newQuantity = 100;
-
-        // when
-        orderItem.updateQuantity(newQuantity);
-
-        // then
-        assertThat(orderItem.getQuantity()).isEqualTo(newQuantity);
-    }
-
-    @Test
-    @DisplayName("주문 수량 변경시 수량이 0 이면 예외가 발생한다")
-    void updateQuantity_Fail_When_Zero_or_Negative() {
-        // given
-        OrderItemCommand command = new OrderItemCommand(productId, quantity);
-        OrderItem orderItem = OrderItem.of(order, command);
-
-        // when & then
-        assertThatThrownBy(() -> orderItem.updateQuantity(0))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("주문 수량은 1개 이상이어야 합니다.");
-
-        assertThatThrownBy(() -> orderItem.updateQuantity(-5))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("주문 수량은 1개 이상이어야 합니다.");
-    }
-}
